+++ main_test.go
@@ -2,10 +2,7 @@ package main
 
 import (
 	"testing"
-
-	"github.com/matryer/is"
 )
 
 func countViolations(t *testing.T) {
-
 }
